+++ examples/with-sse-express/README.md
@@ -1,0 +1,5 @@
+# Muppet SSE Example with Express
+
+This example shows how to use the Muppet with SSE transport on Hono and Express. We are using `SSEServerTransport` transport layer from `@modelcontextprotocol/sdk` to connect.
+
+The `SSEServerTransport` class uses some node specific modules, so it can only be used in a node environment.
