+++ examples/with-sse-express/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "with-sse-nodejs",
+  "name": "with-sse",
   "version": "0.0.1",
   "bin": "./dist/index.js",
   "scripts": {
@@ -7,13 +7,14 @@
     "build": "pkgroll --minify"
   },
   "dependencies": {
-    "@hono/node-server": "^1.14.0",
     "@modelcontextprotocol/sdk": "^1.7.0",
+    "express": "^4.21.2",
     "hono": "^4.7.4",
     "muppet": "workspace:*",
     "zod": "^3.24.2"
   },
   "devDependencies": {
+    "@types/express": "^5.0.1",
     "@types/node": "^22.13.10",
     "pkgroll": "^2.11.2",
     "tsx": "^4.19.3",
