--- examples/with-sse-nodejs/README.md
@@ -1,5 +1,0 @@
-# Muppet SSE Example with Hono
-
-This example shows how to use the Muppet with SSE transport on Hono. We are using `SSEHonoTransport` transport layer from `muppet/streaming` to connect.
-
-This can be used with all the runtimes which supports Streaming with hono. More details about that here - <https://hono.dev/docs/helpers/streaming>
