+++ examples/with-sse/README.md
@@ -1,5 +1,5 @@
-# Muppet SSE Example with Express
+# Muppet SSE Example with Hono
 
-This example shows how to use the Muppet with SSE transport on Hono and Express. We are using `SSEServerTransport` transport layer from `@modelcontextprotocol/sdk` to connect.
+This example shows how to use the Muppet with SSE transport on Hono. We are using `SSEHonoTransport` transport layer from `muppet/streaming` to connect.
 
-The `SSEServerTransport` class uses some node specific modules, so it can only be used in a node environment.
+This can be used with all the runtimes which supports Streaming with hono. More details about that here - <https://hono.dev/docs/helpers/streaming>
